src/core/rag.py
# RAG module - placeholder for Retrieval-Augmented Generation
import faiss
from langchain.vectorstores import FAISS
from langchain.embeddings import HuggingFaceEmbeddings
from sentence_transformers import SentenceTransformer

class RAGHandler:
    def __init__(self):
        # Initialize embeddings
        self.embedding_model = HuggingFaceEmbeddings(model_name="sentence-transformers/all-MiniLM-L6-v2")
        
        # Placeholder for vector store
        self.vectorstore = None

    def load_documents(self, documents):
        # Create vector store from documents
        if documents:
            self.vectorstore = FAISS.from_documents(documents, self.embedding_model)
            return self.vectorstore
        
    def similarity_search(self, query, k=3):
        # Perform similarity search
        if self.vectorstore:
            docs = self.vectorstore.similarity_search(query, k=k)
            return docs
        return []
