requirements.txt
streamlit==1.38.0
transformers==4.41.2
accelerate==0.31.0
torch==2.3.1
sentence-transformers==3.0.1
langchain==0.2.12
langchain-community==0.2.12
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
pandas==2.2.2
gdown==5.2.0
langfuse==2.38.0
python-dotenv
